watermeter

+++ orchid/src/main/java/com/ih2ome/hardware_service/service/model/narcissus/SmartGateway.java
@@ -31,7 +31,7 @@ public class SmartGateway implements Serializable{
   private String remark;
   private Long houseCatalog;
   private Long apartmentId;
-  private Long floor;
+  private int floor;
   private Long houseId;
   private Long roomId;
